User create configuration

+++ app/Http/Controllers/Admin/UserController.php
@@ -6,10 +6,14 @@ use App\Http\Controllers\Controller;
 use App\User;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
+use Spatie\Permission\Models\Role;
 use Yajra\DataTables\DataTables;
 
 class UserController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      *
@@ -34,12 +38,17 @@ class UserController extends Controller
                     return Carbon::parse($row->created_at)->toDateTimeString();
                 })
                 ->addColumn('action', function ($row){
-                    $btn = '<a href="'.url("users/".$row->id."/edit" ).'" class="edit btn btn-warning btn-sm mr-1">Edit</a>';
-                    $btn .= '<a href="javascript:deleteUser('.$row->id.')" class="remove btn btn-danger btn-sm ">Delete</a>';
-                    $btn .= '<form id="deleteForm'.$row->id.'" action="'.url("users/".$row->id).'" method="POST" style="display:none">
+                    if ($row->id == 1){
+                        $btn = '<a href="" class="disabled btn btn-warning btn-sm mr-1" disabled="">Edit</a>';
+                        $btn .= '<a href="" class="disabled btn btn-danger btn-sm " disabled="">Delete</a>';
+                    } else {
+                        $btn = '<a href="'.url("users/".$row->id."/edit" ).'" class="edit btn btn-warning btn-sm mr-1">Edit</a>';
+                        $btn .= '<a href="javascript:deleteUser('.$row->id.')" class="remove btn btn-danger btn-sm ">Delete</a>';
+                        $btn .= '<form id="deleteForm'.$row->id.'" action="'.url("users/".$row->id).'" method="POST" style="display:none">
                                 '.csrf_field().'
                                 '.method_field("DELETE").'
                             </form>';
+                    }
                     return $btn;
                 })
 
@@ -52,22 +61,34 @@ class UserController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\View\View Admin/Users/Create
      */
     public function create()
     {
-        //
+        $roles = Role::all();
+        return view('admin.users.create', compact('roles'));
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param \Illuminate\Http\Request $request
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Validation\ValidationException
      */
     public function store(Request $request)
     {
-        //
+        $validated = $this->validate($request, [
+            'name' => 'required',
+            'email' => 'required|email|unique:users,email',
+            'password' => 'required|same:confirm_password',
+            'roles' => 'required'
+        ]);
+        $validated['password'] = Hash::make($validated['password']);
+        $user = User::create($validated);
+        $user->assignRole($validated['roles']);
+        return redirect()->route('users.index')
+            ->with('success', 'User created successfuly');
     }
 
     /**
@@ -112,6 +133,6 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        dd('2342234');
+        dd($id);
     }
 }
